Add GRE over IPv6 support
Uses the same NLA map as the existing GRE support but changes
`local` and `remote` to type `target`. This further requires the
message stack to have an explicit address family, so logic is added
to pass that back to the link create mechanism if not provided by
the caller. This works for IPRoute, haven't tested IPDB because...

...note that currently it cannot correctly decode iflink messages
from the kernel about v4 or v6 gre tunnels since it doesn't know
what AF those interfaces are in.

This may be a dead-end.